Main page backlink
This is probably not the right syntax, so bear with me @jamesaoverton while I figure these things out. Basically I would like a way to go back to the main page after finishing an edit of purl or metadata file. We probably need to edit this in two different places :D